+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.9.14
 
+* `Ktor`:
+    * Add temporal files uploading functionality (for clients to upload and for server to receive)
+
 ## 0.9.13
 
 * `Versions`: